requirements/requirement.txt
bcrypt==3.2.0
boto3==1.18.16
Django==3.2.6
django-cors-headers==3.7.0
django-crontab==0.7.1
djangorestframework==3.12.4
gunicorn==20.1.0
mysqlclient==2.0.3
PyJWT==2.1.0
Pillow==8.3.1
